feat(getStrata.fn): generate default strata for design-based estimators
pulled code from survey_summary_package::create_strata_fn and added it
to nwfscSurvey as well as expanded the default strata that were available.

Note I did remove one coast_deep, which didn't seem to be used at all in the
code within that package.

Added a strata for sablefish b/c of the specific break at 36.
This function will be used by VASTWestCoast to generate strata for VAST, i.e.,
in the future search for VASTWestCoast::convert_strata4vast.

Thank you @chantelwetzel-NOAA for making the code base for this function
accessible. All I did really was add some default values so that if a number
changes later on, we don't have to look for it in multiple places and made
combinations of strata easier without having to hard code all combinations.