fix(getStrata.fn deep): strata object names cannot have _
The underscore in the strata name that was addded to deep will make
the function look for deep and strat because all names in the current
environment are split using strsplit(..., "_"). The change to the name
happened in 96c38caa1e5964d53a5716c067cc13359d0b22ce, so this commit
reverses this change. There were other changes in that commit as well
so it could not be reversed.