refactor: Updates Roxygen ver, removes Maintainer
CRAN CMD Check was giving error that the
maintainer was not listed in authors.
But, we do not need maintainer given
Chantel's name also has "cre".

https://r-pkgs.org/description.html
An older, still valid approach is to have separate
Maintainer and Author fields in DESCRIPTION.
However, we strongly recommend the more modern approach
of Authors@R and the person() function,
because it offers richer metadata for
various downstream uses.