doc(SexRatio.fn): includes functionality details
Increases the amount of information in the documentation of
SexRatio.fn as understood by KFJ. TODO demarks areas where the
code or the documentation about the code might need changes.
Added verbose = TRUE for a default and a template for verbose
so the documentation can be the same across functions.