fix(plot_westcoast): wrap example in require maps
maps is a suggested package of {ggplot2} and thus not
automatically installed. So, this example will only be ran
when people have the maps package. I did not make it a
requirement, could add to the suggested package list but i
do not know if that will lead to more check errors because
we are not actually using maps only a dependency of it.