fix(plot_comps): Do not limit bubble plotting
limits was eliminating values smaller than 0.1 and larger than 50, so if a
bubble was 51% of total data in a year the bubble was not being plotted. Now,
just remove all zero values so they are not plotted with a small bubble and a
warning message is not printed if ggplot filters the data because of a limited
range. The default of limits is NULL so we don't need the argument any more.

Use a simple method to include zero on the y axis rather than adding
zero to the data. By specifying the lower limit of zero. NA for the
upper limit uses the limit of the data rather than a hard-coded value.
The if else statement allows for lower limit of the data, which is nice
for length composition.