add new function to combine multiple catch records for trawls
Function that can be used if sample types have been used and multiple records are returned for the same trawl_id.  This function would be up to the user to apply but a message has been added to the pull_catch function that prints if trawl_ids are not uniuqe.