remove extra  output from the strata-years ests
This function was also reporting the CV, SE , and log variance but was extremely large when calculated at the strata specific level.  Additionally, the actual upper and lower interval estimates could not be calculated without errors in the code if observations were not available in a strata-year combination limiting the ability to plot these estimates.